zlt initialized to 0, in order to work without MPI communications
This change is introduced to test model with one MPI task using new domain decomposition,
domdec.txt (with one line), but performing the test on the full mediterranean setup, instead of on the testcase
(which has tmask=0 on borders).

Before this change, working on a single task, border values were inizialized to huge(double) but they were overwritten by
mpplnk_my. Deactivation of mpplnk_my results in an overflow.

Initializing zlt=0, we are allowed to work without mpplnk_my, useful to test other branches.